+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -121,7 +121,6 @@ export interface DynamicSettings {
 	distributionMetric: string
 	edgeMetric: string
 	focusedNodePath: string
-	searchedNodePaths: Set<string>
 	searchPattern: string
 	margin: number
 	colorRange: ColorRange