+++ visualization/app/codeCharta/util/dataMocks.ts
@@ -1746,7 +1746,6 @@ export const STATE: State = {
 		distributionMetric: "mcc",
 		edgeMetric: "pairingRate",
 		focusedNodePath: "/root/ParentLeaf",
-		searchedNodePaths: new Set(),
 		searchPattern: "",
 		margin: 48,
 		colorRange: {
@@ -1896,7 +1895,6 @@ export const DEFAULT_STATE: State = {
 		},
 		colorMode: ColorMode.weightedGradient,
 		searchPattern: "",
-		searchedNodePaths: new Set(),
 		sortingOption: SortingOption.NAME,
 		recentFiles: []
 	},